Move check_file_protocols() to utils.py
