+++ satpy/scene.py
@@ -34,6 +34,7 @@ from satpy.dependency_tree import DependencyTree
 from satpy.node import CompositorNode, MissingDependencies, ReaderNode
 from satpy.readers import load_readers
 from satpy.resample import get_area_def, prepare_resampler, resample_dataset
+from satpy.utils import check_file_protocols
 from satpy.writers import load_writer
 
 LOG = logging.getLogger(__name__)
@@ -1504,55 +1505,3 @@ class Scene:
                 LOG.debug("Delayed optional prerequisite for {}: {}".format(comp_id, prereq_id))
 
         return prereq_datasets
-
-
-def check_file_protocols(filenames):
-    """Check filenames for transfer protocols, convert to FSFile objects if possible."""
-    if isinstance(filenames, dict):
-        return _check_file_protocols_for_dicts(filenames)
-    return _check_file_protocols(filenames)
-
-
-def _check_file_protocols_for_dicts(filenames):
-    res = {}
-    for reader, files in filenames.items():
-        res[reader] = _check_file_protocols(files)
-    return res
-
-
-def _check_file_protocols(filenames):
-    local_files, remote_files, fs_files = _sort_files_to_local_remote_and_fsfiles(filenames)
-
-    if remote_files:
-        return local_files + fs_files + _filenames_to_fsfile(remote_files)
-
-    return local_files + fs_files
-
-
-def _sort_files_to_local_remote_and_fsfiles(filenames):
-    from urllib.parse import urlparse
-
-    from satpy.readers import FSFile
-
-    local_files = []
-    remote_files = []
-    fs_files = []
-    for f in filenames:
-        if isinstance(f, FSFile):
-            fs_files.append(f)
-        elif urlparse(f).scheme in ('', 'file') or "\\" in f:
-            local_files.append(f)
-        else:
-            remote_files.append(f)
-    return local_files, remote_files, fs_files
-
-
-def _filenames_to_fsfile(filenames):
-    import fsspec
-
-    from satpy.readers import FSFile
-
-    if filenames:
-        fsspec_files = fsspec.open_files(filenames)
-        return [FSFile(f) for f in fsspec_files]
-    return []
